+++ umongo/dal/motor_asyncio.py
@@ -1,12 +1,12 @@
 import asyncio
-from motor.motor_asyncio import AsyncIOMotorDatabase, AsyncIOMotorCursor
+from motor.motor_asyncio import AsyncIOMotorDatabase, AsyncIOMotorCursor, AsyncIOMotorGridFS
 from pymongo.errors import DuplicateKeyError
 
 from ..abstract import AbstractDal
 from ..data_proxy import DataProxy, missing
-from ..data_objects import Reference
+from ..data_objects import Reference, GridFSReference
 from ..exceptions import NotCreatedError, UpdateError, ValidationError, DeleteError
-from ..fields import ReferenceField, ListField, EmbeddedField
+from ..fields import ReferenceField, ListField, EmbeddedField, GridFSField
 
 from .tools import cook_find_filter
 
@@ -196,6 +196,11 @@ def _reference_io_validate(field, value):
 
 
 @asyncio.coroutine
+def _gridfs_reference_io_validate(field, value):
+    yield from value.exists()
+
+
+@asyncio.coroutine
 def _list_io_validate(field, value):
     validators = field.container.io_validate
     if not validators or not value:
@@ -238,6 +243,9 @@ def _io_validate_patch_schema(schema):
         if isinstance(field, ReferenceField):
             field.io_validate.append(_reference_io_validate)
             field.reference_cls = MotorAsyncIOReference
+        if isinstance(field, GridFSField):
+            field.io_validate.append(_gridfs_reference_io_validate)
+            field.reference_cls = MotorGridFSReference
         if isinstance(field, EmbeddedField):
             field.io_validate.append(_embedded_document_io_validate)
             _io_validate_patch_schema(field.schema)
@@ -261,3 +269,45 @@ class MotorAsyncIOReference(Reference):
                 raise ValidationError(self.error_messages['not_found'].format(
                     document=self.document_cls.__name__))
         return self._document
+
+
+class MotorGridFSReference(GridFSReference):
+
+    def __init__(self, db, collection_name='fs', pk=None):
+        self._gridout = None
+        self._gridfs = None
+        self.db = db
+        self.collection_name = collection_name
+        self.pk = pk
+
+    @property
+    def gridfs(self):
+        if not self._gridfs:
+            self._gridfs = AsyncIOMotorGridFS(self.db, self.collection_name)
+        return self._gridfs
+
+    def fetch(self):
+        """
+        Retrieve from GridFS the referenced file as a GridOut.
+        """
+        if not self._gridout:
+            if self.pk is None:
+                raise ReferenceError('Cannot retrieve a None Reference')
+            self._gridout = yield from self.gridfs.get(self.pk)
+            if not self._gridout:
+                raise ValidationError(self.error_messages['not_found'].format(
+                    gridfs=self.root_collection))
+        return self._gridout
+
+    def exists(self):
+        """
+        Check if the referenced file exists in GridFS.
+        """
+        doc = yield from self.db[self.collection_name].files.find_one(self.pk, fields={'_id': 1})
+        return bool(doc)
+
+    def delete(self):
+        """
+        Delete this file from GridFS.
+        """
+        return self.gridfs.delete(self.pk)